# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "version": "1.0"
    },
    "paths": {
        "/api/getAllCURPS": {
            "get": {
                "summary": "Muestra todos los Curps registrados ",
                "operationId": "c076cbf42fcdafcd045df674fb9e3847",
                "responses": {
                    "200": {
                        "description": "Muestra todos los Curps registrados."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/createCurp": {
            "post": {
                "summary": "Agrega Crups",
                "operationId": "cd821b3c92eb2911f7d8b027468ab850",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellido_p": {
                                        "type": "string"
                                    },
                                    "apellido_m": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string"
                                    },
                                    "sexo": {
                                        "type": "string"
                                    },
                                    "estado": {
                                        "type": "string"
                                    },
                                    "curp": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Bryan Eliut",
                                    "apellido_p": "Hernandez",
                                    "apellido_m": "Moran",
                                    "fecha_nacimiento": "1997-09-04",
                                    "sexo": "H",
                                    "estado": "HG",
                                    "curp": "HEMB970904HHGRRR00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "El Curp se agrego con exito."
                    },
                    "406": {
                        "description": "El request es invalido"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/updateCurp": {
            "put": {
                "summary": "Actualiza los datos del Crup",
                "operationId": "6617d2ba3ffc8a351f07a14de79d81ce",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellido_p": {
                                        "type": "string"
                                    },
                                    "apellido_m": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string"
                                    },
                                    "sexo": {
                                        "type": "string"
                                    },
                                    "estado": {
                                        "type": "string"
                                    },
                                    "curp": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "nombre": "Bryan Eliud",
                                    "apellido_p": "Hernandez",
                                    "apellido_m": "Moran",
                                    "fecha_nacimiento": "1997-09-04",
                                    "sexo": "H",
                                    "estado": "HG",
                                    "curp": "HEMB970904HHGRRR00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "El Curp se agrego con exito."
                    },
                    "406": {
                        "description": "El request es invalido"
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    }
}